@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -33,17 +33,23 @@
  * @codingStandardsIgnoreStart
  */
 
+use _PhpScoper5ea00cc67502b\Mollie\Api\Types\PaymentMethod;
+use _PhpScoper5ea00cc67502b\Mollie\Api\Types\PaymentStatus;
+use Mollie\Repository\PaymentMethodRepository;
+use Mollie\Service\OrderStatusService;
+use PrestaShop\PrestaShop\Adapter\CoreException;
+
 if (!defined('_PS_VERSION_')) {
     exit;
 }
 
-require_once dirname(__FILE__).'/../../mollie.php';
+require_once dirname(__FILE__) . '/../../mollie.php';
 
 /**
  * Class MollieReturnModuleFrontController
  *
  * @property Context? $context
- * @property Mollie   $module
+ * @property Mollie $module
  */
 class MollieReturnModuleFrontController extends ModuleFrontController
 {
@@ -61,15 +67,15 @@ class MollieReturnModuleFrontController extends ModuleFrontController
      * Unset the cart id from cookie if the order exists
      *
      * @throws PrestaShopException
-     * @throws \PrestaShop\PrestaShop\Adapter\CoreException
+     * @throws CoreException
      */
     public function init()
     {
         /** @var Context $context */
         $context = Context::getContext();
         /** @var Cart $cart */
-        $cart = new Cart((int) $this->context->cookie->id_cart);
-        if (Validate::isLoadedObject($cart) && $cart->orderExists()) {
+        $cart = new Cart((int)$this->context->cookie->id_cart);
+        if (Validate::isLoadedObject($cart) && !$cart->orderExists()) {
             unset($context->cart);
             unset($context->cookie->id_cart);
             unset($context->cookie->checkedTOS);
@@ -84,7 +90,7 @@ class MollieReturnModuleFrontController extends ModuleFrontController
      * @throws PrestaShopException
      * @throws Adapter_Exception
      * @throws SmartyException
-     * @throws \PrestaShop\PrestaShop\Adapter\CoreException
+     * @throws CoreException
      */
     public function initContent()
     {
@@ -95,28 +101,30 @@ class MollieReturnModuleFrontController extends ModuleFrontController
 
         parent::initContent();
 
-        $data = array();
+        $data = [];
         $cart = null;
 
+        /** @var PaymentMethodRepository $paymentMethodRepo */
+        $paymentMethodRepo = $this->module->getContainer(PaymentMethodRepository::class);
         /**
          * Set ref is indicative of a payment that is tied to an order instead of a cart, which
          * we still support for transitional reasons.
          */
         if (Tools::getIsset('ref')) {
-            $idOrder = (int) Tools::getValue('id');
+            $idOrder = (int)Tools::getValue('id');
             // Check if user is allowed to be on the return page
-            $data['auth'] = Order::getUniqReferenceOf($idOrder) === Tools::getValue('ref');
+            $data['auth'] = $paymentMethodRepo->getUniqReferenceOf($idOrder) === Tools::getValue('ref');
             if ($data['auth']) {
-                $data['mollie_info'] = Mollie::getPaymentBy('order_id', (int) $idOrder);
+                $data['mollie_info'] = $paymentMethodRepo->getPaymentBy('order_id', (int)$idOrder);
             }
         } elseif (Tools::getIsset('cart_id')) {
-            $idCart = (int) Tools::getValue('cart_id');
+            $idCart = (int)Tools::getValue('cart_id');
 
             // Check if user that's seeing this is the cart-owner
             $cart = new Cart($idCart);
-            $data['auth'] = (int) $cart->id_customer === $this->context->customer->id;
+            $data['auth'] = (int)$cart->id_customer === $this->context->customer->id;
             if ($data['auth']) {
-                $data['mollie_info'] = Mollie::getPaymentBy('cart_id', (int) $idCart);
+                $data['mollie_info'] = $paymentMethodRepo->getPaymentBy('cart_id', (int)$idCart);
             }
         }
 
@@ -124,67 +132,19 @@ class MollieReturnModuleFrontController extends ModuleFrontController
             // any paid payments for this cart?
 
             if ($data['mollie_info'] === false) {
-                $data['mollie_info'] = array();
-                $data['msg_details'] = $this->module->lang('The order with this id does not exist.');
-            } elseif ($data['mollie_info']['method'] === \MollieModule\Mollie\Api\Types\PaymentMethod::BANKTRANSFER
-                && $data['mollie_info']['bank_status'] === \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_OPEN
+                $data['mollie_info'] = [];
+                $data['msg_details'] = $this->l('The order with this id does not exist.');
+            } elseif ($data['mollie_info']['method'] === PaymentMethod::BANKTRANSFER
+                && $data['mollie_info']['bank_status'] === PaymentStatus::STATUS_OPEN
             ) {
-                $data['msg_details'] = $this->module->lang('We have not received a definite payment status. You will be notified as soon as we receive a confirmation of the bank/merchant.');
+                $data['msg_details'] = $this->l('We have not received a definite payment status. You will be notified as soon as we receive a confirmation of the bank/merchant.');
             } else {
-                switch ($data['mollie_info']['bank_status']) {
-                    case 'created':
-                        $data['wait'] = true;
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_OPEN:
-                        $data['wait'] = true;
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_PENDING:
-                        $data['wait'] = true;
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_FAILED:
-                        Tools::redirect($this->context->link->getPagelink('order', true, null, array('step' => 3)));
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_CANCELED:
-                        Tools::redirect($this->context->link->getPagelink('order', true, null, array('step' => 3)));
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_EXPIRED:
-                        $data['msg_details'] = $this->module->lang('Unfortunately your payment was expired.');
-                        break;
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_PAID:
-                    case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_AUTHORIZED:
-                        // Validate the Order
-                        if (isset($cart) && Validate::isLoadedObject($cart)) {
-                            Tools::redirect(
-                                $this->context->link->getPageLink(
-                                    'order-confirmation',
-                                    true,
-                                    null,
-                                    array(
-                                        'id_cart'   => (int) $cart->id,
-                                        'id_module' => (int) $this->module->id,
-                                        'id_order'  => (int) version_compare(_PS_VERSION_, '1.7.1.0', '>=')
-                                            ? Order::getIdByCartId((int) $cart->id)
-                                            : Order::getOrderByCartId((int) $cart->id),
-                                        'key'       => $cart->secure_key,
-                                    )
-                                )
-                            );
-                        }
-
-                        $data['msg_details'] = $this->module->lang('Thank you. Your order has been received.');
-                        break;
-                    default:
-                        $data['msg_details'] = $this->module->lang('The transaction has an unexpected status.');
-                        if (Configuration::get(Mollie::MOLLIE_DEBUG_LOG) == Mollie::DEBUG_LOG_ERRORS) {
-                            Logger::addLog(__METHOD__.'said: The transaction has an unexpected status ('.$data['mollie_info']['bank_status'].')', Mollie::WARNING);
-                        }
-                        break;
-                }
+                $data['wait'] = true;
             }
         } else {
             // Not allowed? Don't make query but redirect.
-            $data['mollie_info'] = array();
-            $data['msg_details'] = $this->module->lang('You are not authorised to see this page.');
+            $data['mollie_info'] = [];
+            $data['msg_details'] = $this->l('You are not authorised to see this page.');
             Tools::redirect(Context::getContext()->link->getPageLink('index', true));
         }
 
@@ -197,7 +157,7 @@ class MollieReturnModuleFrontController extends ModuleFrontController
                 $this->context->link->getModuleLink(
                     $this->module->name,
                     'return',
-                    array('ajax' => 1, 'action' => 'getStatus', 'transaction_id' => $data['mollie_info']['transaction_id']),
+                    ['ajax' => 1, 'action' => 'getStatus', 'transaction_id' => $data['mollie_info']['transaction_id']],
                     true
                 )
             );
@@ -210,15 +170,15 @@ class MollieReturnModuleFrontController extends ModuleFrontController
     /**
      * Prepend module path if PS version >= 1.7
      *
-     * @param string      $template
-     * @param array       $params
+     * @param string $template
+     * @param array $params
      * @param string|null $locale
      *
      * @throws PrestaShopException
      *
      * @since 3.3.2
      */
-    public function setTemplate($template, $params = array(), $locale = null)
+    public function setTemplate($template, $params = [], $locale = null)
     {
         if (version_compare(_PS_VERSION_, '1.7.0.0', '>=')) {
             $template = "module:mollie/views/templates/front/17_{$template}";
@@ -256,7 +216,7 @@ class MollieReturnModuleFrontController extends ModuleFrontController
      *
      * @throws PrestaShopException
      * @throws Adapter_Exception
-     * @throws \PrestaShop\PrestaShop\Adapter\CoreException
+     * @throws CoreException
      * @throws SmartyException
      *
      * @since 3.3.2
@@ -264,97 +224,150 @@ class MollieReturnModuleFrontController extends ModuleFrontController
     protected function processGetStatus()
     {
         header('Content-Type: application/json;charset=UTF-8');
+        /** @var PaymentMethodRepository $paymentMethodRepo */
+        $paymentMethodRepo = $this->module->getContainer(PaymentMethodRepository::class);
 
         $transactionId = Tools::getValue('transaction_id');
-        $dbPayment = Mollie::getPaymentBy('transaction_id', $transactionId);
+        $dbPayment = $paymentMethodRepo->getPaymentBy('transaction_id', $transactionId);
         $cart = new Cart($dbPayment['cart_id']);
         if (!Validate::isLoadedObject($cart)) {
-            die(json_encode(array(
+            die(json_encode([
                 'success' => false,
-            )));
+            ]));
         }
-        $orderId = (int) version_compare(_PS_VERSION_, '1.7.1.0', '>=')
-            ? Order::getIdByCartId((int) $cart->id)
-            : Order::getOrderByCartId((int) $cart->id);
-        $order = new Order((int) $orderId);
-        if (Validate::isLoadedObject($order)) {
-            die(json_encode(array(
-                'success'  => true,
-                'status'   => static::DONE,
-                'response' => null,
-                'href'     => $this->context->link->getPageLink(
-                    'order-confirmation',
-                    true,
-                    null,
-                    array(
-                        'id_cart'   => (int) $cart->id,
-                        'id_module' => (int) $this->module->id,
-                        'id_order'  => (int) $order->id,
-                        'key'       => $cart->secure_key,
-                    )
-                )
-            )));
-        }
+        $orderId = Order::getOrderByCartId((int)$cart->id);
+        $order = new Order((int)$orderId);
+
         if (!Validate::isLoadedObject($cart)) {
-            die(json_encode(array(
+            die(json_encode([
                 'success' => false,
-            )));
+            ]));
         }
 
-        if ((int) $cart->id_customer !== (int) $this->context->customer->id) {
-            die(json_encode(array(
+        if ((int)$cart->id_customer !== (int)$this->context->customer->id) {
+            die(json_encode([
                 'success' => false,
-            )));
+            ]));
         }
 
 
         if (!Tools::isSubmit('module')) {
             $_GET['module'] = $this->module->name;
         }
-        $webhookController = new MollieWebhookModuleFrontController();
-        if (Tools::substr($transactionId, 0, 3) === 'ord') {
-            $apiPayment = $webhookController->processTransaction($this->module->api->orders->get($transactionId, array('embed' => 'payments')));
-        } else {
-            $apiPayment = $webhookController->processTransaction($this->module->api->payments->get($transactionId));
-        }
 
-        switch ($apiPayment->status) {
-            case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_EXPIRED:
-            case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_FAILED:
-            case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_CANCELED:
+        if (Tools::substr($transactionId, 0, 3) === 'ord') {
+            $transaction = $this->module->api->orders->get($transactionId, ['embed' => 'payments']);
+        } else {
+            $transaction = $this->module->api->payments->get($transactionId);
+        }
+        /** @var OrderStatusService $orderStatusService */
+        $orderStatusService = $this->module->getContainer(OrderStatusService::class);
+
+        $orderStatus = $transaction->status;
+        switch ($transaction->status) {
+            case PaymentStatus::STATUS_EXPIRED:
+            case PaymentStatus::STATUS_FAILED:
+            case PaymentStatus::STATUS_CANCELED:
+                $orderStatus = $transaction->status;
+                $orderStatusId = (int)Mollie\Config\Config::getStatuses()[$orderStatus];
+
+                $orderStatusService->setOrderStatus($orderId, $orderStatusId);
+
+                $failUrl = $this->context->link->getModuleLink(
+                    $this->module->name,
+                    'fail',
+                    [
+                        'cartId' => $cart->id,
+                        'secureKey' => $cart->secure_key,
+                        'orderId' => $orderId,
+                        'moduleId' => $this->module->name,
+                    ],
+                    true
+                );
+
+                die(json_encode([
+                    'success' => true,
+                    'status' => static::DONE,
+                    'response' => json_encode($transaction),
+                    'href' => $failUrl
+
+                ]));
+            case PaymentStatus::STATUS_AUTHORIZED:
+            case PaymentStatus::STATUS_PAID:
                 $status = static::DONE;
-            die(json_encode(array(
-                'success'  => true,
-                'status'   => $status,
-                'response' => json_encode($apiPayment),
-                'href'     => $this->context->link->getPagelink('order', true, null, array('step' => 3))
-            )));
-            case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_AUTHORIZED:
-            case \MollieModule\Mollie\Api\Types\PaymentStatus::STATUS_PAID:
-                $status = static::DONE;
+                $orderDetails = $order->getOrderDetailList();
+                /** @var OrderDetail $detail */
+                foreach ($orderDetails as $detail) {
+                    $orderDetail = new OrderDetail($detail['id_order_detail']);
+                    if (
+                        Configuration::get('PS_STOCK_MANAGEMENT') &&
+                        ($orderDetail->getStockState() || $orderDetail->product_quantity_in_stock < 0)
+                    ) {
+                        $orderStatus = Mollie\Config\Config::STATUS_PAID_ON_BACKORDER;
+                        break;
+                    }
+                }
                 break;
             default:
                 $status = static::PENDING;
+                $orderDetails = $order->getOrderDetailList();
+                /** @var OrderDetail $detail */
+                foreach ($orderDetails as $detail) {
+                    $orderDetail = new OrderDetail($detail['id_order_detail']);
+                    if (
+                        Configuration::get('PS_STOCK_MANAGEMENT') &&
+                        ($orderDetail->getStockState() || $orderDetail->product_quantity_in_stock < 0)
+                    ) {
+                        $orderStatus = Mollie\Config\Config::STATUS_PENDING_ON_BACKORDER;
+                        break;
+                    }
+                }
                 break;
         }
 
-        die(json_encode(array(
-            'success'  => true,
-            'status'   => $status,
-            'response' => json_encode($apiPayment),
-            'href'     => $this->context->link->getPageLink(
-                'order-confirmation',
-                true,
-                null,
+        $orderStatusId = (int)Mollie\Config\Config::getStatuses()[$orderStatus];
+        $orderStatusService->setOrderStatus($orderId, $orderStatusId);
+        $this->savePaymentStatus($transactionId, $orderStatus, $orderId);
+
+        $successUrl = $this->context->link->getPageLink(
+            'order-confirmation',
+            true,
+            null,
+            [
+                'id_cart' => (int)$cart->id,
+                'id_module' => (int)$this->module->id,
+                'id_order' => (int)version_compare(_PS_VERSION_, '1.7.1.0', '>=')
+                    ? Order::getIdByCartId((int)$cart->id)
+                    : Order::getOrderByCartId((int)$cart->id),
+                'key' => $cart->secure_key,
+            ]
+        );
+
+        die(json_encode([
+            'success' => true,
+            'status' => $status,
+            'response' => json_encode($transaction),
+            'href' => $successUrl
+        ]));
+    }
+
+    protected function savePaymentStatus($transactionId, $status, $orderId)
+    {
+        try {
+            return Db::getInstance()->update(
+                'mollie_payments',
                 array(
-                    'id_cart'   => (int) $cart->id,
-                    'id_module' => (int) $this->module->id,
-                    'id_order'  => (int) version_compare(_PS_VERSION_, '1.7.1.0', '>=')
-                        ? Order::getIdByCartId((int) $cart->id)
-                        : Order::getOrderByCartId((int) $cart->id),
-                    'key'       => $cart->secure_key,
-                )
-            )
-        )));
+                    'updated_at'  => array('type' => 'sql', 'value' => 'NOW()'),
+                    'bank_status' => pSQL($status),
+                    'order_id'    => (int) $orderId,
+                ),
+                '`transaction_id` = \''.pSQL($transactionId).'\''
+            );
+        } catch (Exception $e) {
+            /** @var PaymentMethodRepository $paymentMethodRepo */
+            $paymentMethodRepo = $this->module->getContainer(PaymentMethodRepository::class);
+            $paymentMethodRepo->tryAddOrderReferenceColumn();
+            throw $e;
+        }
     }
 }
