@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -64,6 +64,7 @@ const SortableList = SortableContainer(({ items, translations, onArrowClicked, o
             code={item.id}
             enabled={item.enabled}
             available={item.available}
+            tipEnableSSL={item.tipEnableSSL}
             translations={translations}
             position={index}
             max={items.length - 1}
