@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -30,53 +30,56 @@
  * @package    Mollie
  * @link       https://www.mollie.nl
  */
-import React, { ReactElement, useCallback } from 'react';
-import { useMappedState } from 'redux-react-hook';
+import React, {ReactElement, useCallback} from 'react';
+import {useMappedState} from 'redux-react-hook';
 
 import OrderPanelContent from '@transaction/components/orderlines/OrderPanelContent';
+import WarningContent from '@transaction/components/orderlines/WarningContent';
 
 export default function OrderPanel(): ReactElement<{}> {
-  const { config: { legacy, moduleDir }, config } = useCallback(useMappedState((state: IMollieOrderState): any => ({
-    translations: state.translations,
-    config: state.config,
-    order: state.order,
-  })), []);
+    const {config: {legacy, moduleDir}, config} = useCallback(useMappedState((state: IMollieOrderState): any => ({
+        translations: state.translations,
+        config: state.config,
+        order: state.order,
+    })), []);
 
-  if (Object.keys(config).length <= 0) {
-    return null;
-  }
+    if (Object.keys(config).length <= 0) {
+        return null;
+    }
 
-  if (legacy) {
+    if (legacy) {
+        return (
+            <fieldset style={{marginTop: '14px'}}>
+                <legend>
+                    <img
+                        src={`${moduleDir}views/img/logo_small.png`}
+                        width="32"
+                        height="32"
+                        alt="Mollie logo"
+                        style={{height: '16px', width: '16px', opacity: 0.8}}
+                    />
+                    &nbsp;<span>Mollie</span>&nbsp;
+                </legend>
+                <WarningContent/>
+                <OrderPanelContent/>
+            </fieldset>
+        );
+    }
+
     return (
-      <fieldset style={{ marginTop: '14px' }}>
-        <legend>
-          <img
-            src={`${moduleDir}views/img/logo_small.png`}
-            width="32"
-            height="32"
-            alt="Mollie logo"
-            style={{ height: '16px', width: '16px', opacity: 0.8 }}
-          />
-          &nbsp;<span>Mollie</span>&nbsp;
-        </legend>
-        <OrderPanelContent/>
-      </fieldset>
+        <div className="panel">
+            <div className="panel-heading">
+                <img
+                    src={`${moduleDir}views/img/mollie_panel_icon.png`}
+                    width="32"
+                    height="32"
+                    alt="Mollie logo"
+                    style={{height: '16px', width: '16px', opacity: 0.8}}
+                />
+                &nbsp;<span>Mollie</span>&nbsp;
+            </div>
+            <WarningContent/>
+            <OrderPanelContent/>
+        </div>
     );
-  }
-
-  return (
-    <div className="panel">
-      <div className="panel-heading">
-        <img
-          src={`${moduleDir}views/img/mollie_panel_icon.png`}
-          width="32"
-          height="32"
-          alt="Mollie logo"
-          style={{ height: '16px', width: '16px', opacity: 0.8 }}
-        />
-        &nbsp;<span>Mollie</span>&nbsp;
-      </div>
-      <OrderPanelContent/>
-    </div>
-  );
 }
