@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -38,12 +38,12 @@ import { useMappedState } from 'redux-react-hook';
 import { formatCurrency } from '@shared/tools';
 
 export default function PaymentInfoContent(): ReactElement<{}> {
-  const { translations, payment, currencies, config: { legacy } }: Partial<IMollieOrderState> = useCallback(useMappedState((state: IMollieOrderState): any => ({
+  const { translations, payment, currencies, config: { legacy } }: Partial<IMollieOrderState> = useMappedState((state: IMollieOrderState): any => ({
     payment: state.payment,
     currencies: state.currencies,
     translations: state.translations,
     config: state.config,
-  })), []);
+  }));
 
   return (
     <>
