@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -45,11 +45,11 @@ const Div = styled.div`
 ` as any;
 
 export default function RefundInfo(): ReactElement<{}> {
-  const { translations, config: { legacy }, payment }: Partial<IMollieOrderState> = useCallback(useMappedState( (state: IMollieOrderState): any => ({
+  const { translations, config: { legacy }, payment }: Partial<IMollieOrderState> = useMappedState( (state: IMollieOrderState): any => ({
     payment: state.payment,
     translations: state.translations,
     config: state.config,
-  })), []);
+  }));
 
   if (legacy) {
     return (
