@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2012-2019, Mollie B.V.
+ * Copyright (c) 2012-2020, Mollie B.V.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -36,12 +36,13 @@ import { useMappedState } from 'redux-react-hook';
 import PaymentInfo from '@transaction/components/refund/PaymentInfo';
 import RefundInfo from '@transaction/components/refund/RefundInfo';
 import LoadingDots from '@shared/components/LoadingDots';
+import WarningContent from "@transaction/components/orderlines/WarningContent";
 
 export default function RefundPanel(): ReactElement<{}> {
-  const { payment, config }: Partial<IMollieOrderState> = useCallback(useMappedState((state: IMollieOrderState): any => ({
+  const { payment, config }: Partial<IMollieOrderState> = useMappedState((state: IMollieOrderState): any => ({
     config: state.config,
     payment: state.payment,
-  }),), []);
+  }),);
 
   if (Object.keys(config).length <= 0) {
     return null;
@@ -61,6 +62,7 @@ export default function RefundPanel(): ReactElement<{}> {
           />
           <span>Mollie</span>&nbsp;
         </legend>
+        <WarningContent/>
         {!payment && <LoadingDots/>}
         {!!payment && payment.status && (
           <>
@@ -83,6 +85,7 @@ export default function RefundPanel(): ReactElement<{}> {
           style={{ height: '16px', width: '16px', opacity: 0.8 }}
         /> <span>Mollie</span>&nbsp;
       </div>
+      <WarningContent/>
       {!payment && <LoadingDots/>}
       {!!payment && payment.status && (
         <div className="panel-body row">
